Added in testing program, and Updated version of library
This version of the library requires the user to .Close() the ExcelReader object when they are done using it. However, this runs significantly faster on the order of about 10 times faster than previous versions.

I was having an issue with pop-ups and lingering background processes but it seems that I have (for the most part) eliminated these. One pop-up, with what I assume is an attached/sibling background process remains, but I believe this may in fact be connected to the fact that my windows is not activated, thus other users shouldn't have that issue.